tests - delete table test refactor
